backend_flask/requirements.txt
flask-cors==6.0.2
flask-migrate==4.1.0
flask-socketio==5.6.1
gevent==25.9.1
importlib-metadata==8.7.1
jaraco-context==6.1.0
jaraco-functools==4.4.0
jaraco.text==4.0.0
lap==0.5.13
openvino==2026.0.0
pip-chill==1.0.3
platformdirs==4.4.0
pymysql==1.1.2
python-dotenv==1.2.2
tomli==2.4.0
ultralytics==8.4.21
zope-event==6.1
zope-interface==8.2
cryptography==42.0.5
